requirements.txt
dnspython
requests
setuptools
pyfiglet